instrument code in CSL debug builds to record sizes of vectors allocated
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@1886 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ scripts/hello.c
@@ -2,6 +2,6 @@
 
 int main(int argc, char *argv[])
 {
-    printf("Hello, sizeof(void *) = %d\n", sizeof(void *));
+    printf("Hello, sizeof(void *) = %d\n", (int)sizeof(void *));
     return 0;
 }
